+++ db/psql_schema/0.9/007.multilingual.sql
@@ -1,1 +1,0 @@
-ALTER TABLE users ADD COLUMN preferred_language text NOT NULL DEFAULT 'ro';